[eshell] Protect prompt only
If the user enters text while a command is running and the command does not
consume the text, eshell will input this text after it displays the prompt.
Unfortunately, that would happen after this function was called and this text
would end up getting protected.

This change protects up until the end of the last output (the end of the
prompt).